requirements.txt
Django==2.2.2
djangorestframework==3.10.3
Pillow==6.0.0
psycopg2-binary==2.8.3
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
